# src/utils/swagger.yaml
swagger: '2.0'
info:
  description: 'Endpoint Swagger of RzBlog'
  version: '1.0.0'
  title: 'Swagger RzBlog 2.0'
  termsOfService: 'http://127.0.0.1:5000/api'
  contact:
    email: 'rzmobiledev@gmail.com'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: '127.0.0.1:5000'
basePath: '/api'
...
schemes:
  - 'http'
  - 'https'
securityDefinitions:
  api_key:
    type: 'apiKey'
    name: 'Authorization'
    in: 'header'
paths:
...
  /article/{id}:
    get:
      tags:
        - 'Article'
      summary: 'List article by article id'
      description: 'All unauthenticated user can perform this action'
      operationId: 'listOneArticle'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'Article id'
          required: true
          type: 'integer'
      responses:
        '200':
          description: 'OK'
          schema:
            $ref: '#/definitions/Article'
        '400':
          description: 'Unknown errors'
        '404':
          description: 'Article not found.'

    put:
      tags:
        - 'Article'
      summary: 'Edit article by article id.'
      description: 'User need authentication credentials to access endpoint.'
      operationId: 'updateArticle'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'Article id'
          required: true
          type: 'integer'
        - in: 'body'
          name: 'body'
          description: 'Article object payload that needs to be added to the database'
          required: true
          schema:
            $ref: '#/definitions/Article'
      responses:
        '200':
          description: 'Ok'
          schema:
            $ref: '#/definitions/Article'
        '400':
          description: 'Unknown errors'
        '404':
          description: 'Article not found.'
      security:
        - api_key: []

    delete:
      tags:
        - 'Article'
      summary: 'delete article by article id.'
      description: 'All unauthenticated user can perform this action'
      operationId: 'deleteArticle'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'Article id'
          required: true
          type: 'integer'
      responses:
        '200':
          description: 'Article deleted'
        '400':
          description: 'Unknown errors'
        '401':
          description: 'Access Denied!'
        '404':
          description: 'Article not found.'
      security:
        - api_key: []
...
  /category/{id}:
    get:
      tags:
        - 'Category'
      summary: 'Retrieve a category'
      description: 'Show one category based on id selected'
      operationId: 'getOneCategory'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'Category id'
          required: true
          type: 'integer'
      responses:
        '200':
          description: 'Ok'
          schema:
            $ref: '#/definitions/Category'
        '400':
          description: 'Unknown errors.'
        '404':
          description: 'Category not found'
    put:
      tags:
        - 'Category'
      summary: 'Update a category'
      description: ''
      operationId: 'updateCategory'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'Category id'
          required: true
          type: 'integer'
        - in: 'body'
          name: 'body'
          description: 'Category object payload'
          required: true
          schema:
            $ref: '#/definitions/Category'
      responses:
        '200':
          description: 'Ok'
          schema:
            $ref: '#/definitions/Category'
        '400':
          description: 'Unknown errors.'
        '401':
          description: 'Unauthorized.'
        '403':
          description: 'Access denied'
        '404':
          description: 'Category not found.'
      security:
        - api_key: []

    delete:
      tags:
        - 'Category'
      summary: 'Delete a category'
      description: ''
      operationId: 'deleteCategory'
      consumes:
        - 'application/json'
        - 'application/xml'
      produces:
        - 'application/json'
        - 'application/xml'
      parameters:
        - name: 'id'
          in: 'path'
          description: 'Category id'
          required: true
          type: 'integer'
      responses:
        '200':
          description: 'Category deleted.'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Category not found.'
        '403':
          description: 'Access denied'
      security:
        - api_key: []
...
definitions:
...
  Article:
    type: 'object'
    properties:
      categoryId:
        type: 'integer'
        format: 'int64'
        required: true
      title:
        type: 'string'
        description: 'Title of article'
      subtitle:
        type: 'string'
        description: 'Subtitle of article'
      description:
        type:: 'string'
        description: 'Content of articles'
      tags:
        type: 'array'
        items:
          $ref: '#/definitions/Tag'
  Tag:
    type: 'object'
    properties:
      name:
        type: 'string'
  Category:
    type: 'object'
    required:
      - 'name'
    properties:
      name:
        type: 'string'
        description: 'Name of category'
...